add icons to intro
