--- src/components/home-page/introduction/headShotWithText.module.css
@@ -1,59 +1,0 @@
-.main{
-  padding-top: 54px;
-  display:block;
-  @media(min-width: 720px){
-    display:flex;
-    border-bottom: solid 1px #99CC00;
-    margin-bottom:18px;
-  }
-}
-.img_headshot{
-  background-color:black;
-  z-index: 2;
-  width:100%;
-  border-bottom:1px solid #99cc00;
-  @media(min-width: 720px){
-    width: 350px;
-    height: 300px;
-    border-bottom:none;
-    background-color:black;
-  }
-}
-
-.container_text{
-padding:18px 18px 0 18px;
-display:flex;
-justify-content: center;
-flex-direction: column;
-background-color:black;
-color:#99cc00;
-@media(min-width: 720px){
-  color: #99cc00;
-  background-color:black;
-}
-}
-.container_text p{
-  @media(max-width:720px){
-    margin:0;
-  }
-}
-
-.container_title_text{
-  background-color:black;
-  @media(min-width:720px){
-    display:none;
-  }
-  text-align:center;
-  padding:18px;
-  h4{
-    margin:0;
-    color: #99cc00;
-  }
-
-}
-.desktop_title{
-  display:none;
-  @media(min-width: 720px){
-    display:block
-  }
-}
