+++ src/components/render-icon/RenderIcon.js
@@ -1,14 +1,9 @@
 import React from 'react';
 
-
-
-
-export default function RenderIcon({ iconArray, direction }) {
-
-  
+export default function RenderIcon({ iconArray, direction, color='white' }) {
   const iconList = iconArray.map((icon, i) => {
     if (icon.link && icon.text && icon.icon) {
-      return <li key={i} css={{ alignSelf:'center', color: 'white', margin:'1rem'}}>
+      return <li key={i} css={{ alignSelf:'center', color: color, margin:'1rem'}}>
         <a css={{ display: 'flex' }} href={icon.link}>
           <div css={{ display: 'flex', justifyContent: 'center' }}>
             <p css={{ alignSelf: 'center', margin: '0 10px 0 0' }}>{icon.icon}</p>
@@ -19,7 +14,7 @@ export default function RenderIcon({ iconArray, direction }) {
       </li>
 
     } else if (icon.text && icon.icon) {
-      return <li key={i} css={{ alignSelf: 'center', color: 'white', margin:'1rem'}}>
+      return <li key={i} css={{ alignSelf: 'center', color: color, margin:'1rem'}}>
         <div css={{ display: 'flex', justifyContent: 'center' }}>
           <p css={{ alignSelf: 'center', margin: '0 10px 0 0' }}>{icon.icon}</p>
           <p css={{ alignSelf: 'center', margin: '0' }}>{icon.text}</p>
@@ -27,7 +22,7 @@ export default function RenderIcon({ iconArray, direction }) {
         </div>
       </li>
     }
-    else return <li key={i} css={{ alignSelf: 'center', color: 'white', fontSize: '32px', margin:'1rem'}}>
+    else return <li key={i} css={{ alignSelf: 'center', color: color, fontSize: '32px', margin:'1rem'}}>
       <div css={{display:'flex'}}>
         <p css={{ alignSelf: 'center', margin: '0 10px 0 0' }}>{icon.icon}</p>
       </div>
