+++ src/content/aboutMe.js
@@ -1,5 +1,5 @@
 export const title = 'JavaScript Developer | Perpetual Tinkerer'
 
-export const aboutMe = `I'm Olli, a JavaScript Developer and a perpetual tinkerer with a
+export const aboutMe = `I'm Olli, a coder with a
 passion for diving underthe hood to get a deeper understanding of new technology. My ability to contribue is very important to me, so please let me know
 if I can help you in any way, or if you have any questions.`
